Initial cut at a controllable character (#5)
A few project level structural changes and then an initial cut at a character component.

Structural

    Project settings: Enabled untyped declarations as an error. This means we need to write well-typed gdscript. In addition to being easier to reason about we get a meaningful perf boost. It's not c# but it's better than nothing.
    Physics layer meanings
        Layer 1: Allocated for collisions
        Layer 2: Allocated for player/world interaction hotspots
    Input map additions: up/down/left/right
    Initial rough project directories.
        Art/ — you guessed it
        Audio/ — what it says on the box
        Scenes/ — yup
        Scrap/ — Things that are not intended to live to release. Things that go here are intended to integration test mechanics or functionality of other work
        Scripts/ — code that isn't tightly bound to a corresponding scene/component.

Code changes
Character

This scene is intended to be a base for any character that moves -- it has slots for per-instantiation override of the animated sprite 2d associated with movement and works in an 8-directional regime. If need be we can reduce this to 4 with minimal effort.

In order to set this as a player-controlled character there's a toggle you can enable when adding Devin to a scene. Nothing is done to support attacking yet since we don't know how that works/if that will exist.
Enums

This is intended to be almost exclusively a way to make type-safe hard-coded / non user-facing values and have a canonical value -> string mapping (if needed). Other likely uses will be states for any state machines we have.
Utils

This will possibly be contentious as Utils is generally a trash pattern that collections random unrelated functions and results in a structure that's hard to reason about. It's plausible we will want to eventually remove everything from here...

But for now I'm leaving it in as a starting point to put code that is potentially usable beyond the initial need that resulted in it being written. So for now "angle to direction" seems potentially relevant to a couple of use cases as we build out mechanics so here it goes