Adds detectable component proof of concept (#9)
Summary

    Switches light to a 360 degree pattern based on feedback from Keumars
    Adds scene we can add to parent that takes action when detected/"undetected" by the lamp. Currently used to hide and show the two special trees

Details

The detectable scene has two modes:

    if no connections are made to the detected signal it does a default of get_parent().show() or hide() to make the simple thing (make things visible or not) easy
    if a connection is made to the detected signal it will emit that with a bool indicating if the object was detected or undetected.

Lamp has been updated so that each lamp light level has a corresponding beam sensor (an Area2D with appropriate mask set). When switching between lamp levels the non-active level sensors will be set to a) .enabled = false to stop emitting light and b) the level's Area2D with get .monitoring = false to stop detecting items.
Detection

When a Detectable passes into the Area2D of the lamp the corresponding Lamp's signal handler will get triggered which calls set_detected kicking off the configured effect.

We use this setup (instead of the Detectables being the initiator of action) because it felt logically sound that the lamp would be the thing that triggers action. It would be plausible to switch the control here but it's not clear there is a need.
Demo

    A demo of the old (beam-based) lamp/detection is available here -- https://youtu.be/d_KSZtceeB0
    A newer gif of radial detection is available in this thread