Level Design / graybox initial (#11)
Expanded the map on the TestMovement scene, set that to the default, and did most of the in-engine work on that scene. I added a few png files for visual placeholders and colliders where necessary. In addition to initial grayboxing, I also added WASD support and a zoom via the mousewheel so I could expand the map in real-time.