Adding an "Interactable" node + some main skeleton (#14)
Summary

This does some initial setup for a TGO main scene, menus, overlays, and debug launchers.

Additionally it adds a new scene/node called Interactable that can be added allowing for user-driven or automatic signals to be emitted when a character is in the area.

And finally we get a new debugging scene where I test some physics stuff to make sure that Devin can exist as a peer to his world (which should make level transfer easier) and the aforementioned interactable plumbing.
Interaction

New Input Event: A new input event has been added and can be triggered via , Enter, or the A-key on a controller (& iirc Cross for PS controller).

Behavior: A node is marked interactable by adding an Interactable child and assigning it a CollisionShape2D. It will then fire a signal passing along the Character that triggered the interaction.

Additionally an interaction area may be marked as "automatic" which means it will fire as soon as the character touches it with their interaction sensor instead of waiting for user to hit a key.

Future: plenty of default behaviors need to be added on top of this, e.g., we need Item/NPC/World objects (doors, chests, etc) variations. We'll also probably want some kind of UI toast.

Caveat: This works based on a sensor that runs ahead of a character in the direction they're facing and does not detect when the character enters a space. This means something like a "proximity grenade" or whatever might be a little weird wrt when it triggers but it could be made to work with an automatic trigger. It's plausible we could rewrite this to act with inverted behavior, e.g., the event area acts as the sensing device and watches for a Character to "enter" it but we can cross that bridge when we get there / if we can't get this to work as we want. Broadly speaking I think this will completely suffice for us.

Structure

Loosely follows the structured described here

Introduces main scene Driver that:

    bootstraps the initial scene, currently just a Debug menu but we'll eventually switch that up
    has an API to load a debug scene. This is probably similar-to but not quite the same pattern we'll use for other menu callbacks. But it does respect the signal-up/call-down philosophy while not digging into its children's structure...
    with the exception of curtain which may or may not stick around as currently set up, TBD

Menus

Adds a DebugMenu in place that can be used to select a test scene to load. It's the only place we should be calling into ZZ_Scratch from "prod" code. The contents of the debug menu are collected by calling into ScratchScenes.get_scenes which is documented wrt behavior.
Managers

AudioManager, OverlayManager, and MenuManager exist as stubs and are not fleshed out as we don't yet know enough about their behavior or the scenes that will live under them to do high-confidence prototying.