Introduce top level AudioManager (#16)
Summary

Introduces an AudioManager that lives at the top level and configures an initial set of audio busses.

Currently basic functionality only; we intend to transition this to middleware and I don't want to overengineer what we can do here when it'll likely be eclipsed out of the box by wwise / fmod. Anyway, this handles:

    Play a specific track as background music
    crossfade between tracks if the one you request isn't playing but something else is
    stop with/without fade
    a [0, 1] scale that is detached from the DB scale used by godot (more about this later) per bus
    saving/loading bus volume levels to a prefs file

Git LFS

This change also starts using git lfs to store mp3 files, see git-lfs docs to learn how to set it up. If everything is working correctly it's usually a single command git lfs install
Decibels vs Percentage

So the best practice seems to be "use db!" but in most circumstances -50 is the same as -80 is essentially the same as -35 etc.

So I've introduced a DB_MIN/DB_MAX that we use to interpret volume levels set on a 0 -> 1 scale. Basically if you build a UI that has a slider you can set it to 0 and you'll only get DB_MIN. This makes the slider more useful because in normal usage you can use much more of it to control the volume instead of the top 10%.
Bus vs Player volume

One other thing here is that we're using bus levels to control volume instead of player volume. My driving reason here is twofold:

    It's more representative of what we want when we change volume. That isn't as big a deal when there is primarily one player, e.g., background music but it's more impactful/useful when dealing with things that may have several players like sound effects or menu effects.
    It's less of a headache to save/load level values because there are fewer nodes we need to track and they transcend scene boundaries in an expected way. That is: if i turn down menu noises in the Options menu I also expect menu noises in game to be reduced.

Honestly if you squint those two are kind of the same-but-different reason 🤷
Demo scene

Not much to say here -- I added a new scratch demo scene to play with features. It led directly to adding some safe guards to bus pct -> volume_db mappings when I had forgot a - and accidentally set background music to +50db. Bad times.