[large] Add level loading, greybox infrastructure, dialogue interactable (#19)
## Summary

Introduces a prospective level structure for use in grey boxing work as well as some tooling to make it approachable to level build for designers. This is a first pass and we'll continue to add more as time goes on based on design request and intuition.

## Details: LevelBase / level loading

The best summary of this exists in [the docs](https://github.com/Small-Loan-Studio/TGO/blob/envy-level-stubs/docs/tgo-levels.md) of the design. What follows is a tl;dr:

1. Add a tilemap with default layers and some opinions on how they'll be used
2. Declare by fiat that most dynamic/interactable objects will exist outside of the tilemap and allocate a 
3. Add a marker system that we'll use as player destinations when moving players around the world
4. Expand the Driver codebase to handle:
   a. loading new scenes (levels) upon request; this includes a hook to perform any setup if it needs to reload any persistent state
   b. teleporting the player to a destination in that level (c.f. marker system)
   c. unloading the previously loaded scene with hooks for that level if it needs to save any state for future loads
5. Adds a new InteractableAction that can trigger a level load/player move

## Details: Dialogic / InteractableDialogue
There's two new scratch scene that triggers lets us work through some proof of life stuff with the dialogic addon. This is paired with a new `InteractableDialogue` action that can be used to trigger a DL timeline when interacted with.

It doesn't currently let you jump to labels but that is a pretty trivial addition if necessary.

## Details: Greyboxing
Adds a new object called `GreyboxObject` that (for now) we plan to primarily use for level layout.

The base scene is a custom deal that makes it easy to set collision/interactable state via the editor so you don't have to reconstruct the pile of bullshit (root, sprite2d, area2d, interactable, collisionshape2ds) every time you want to add something to the game.

It also includes some editor-only debugging interface to mark the y-sort line so that you can tell where the "is behind object" point will be and adjust accordingly.

Demo of the greybox stuff is here  https://www.youtube.com/watch?v=iI7_zmBksH8

## Details: Interactable
Not much changed here! Mostly they now have a reference to the level in which they exists. Currently this is used primarily for plumbing a `Driver` reference so that we can request level changes but eventually I could see this as a path to get quest/inventory/character/etc references.


## Bonus changes

- Some messing around with the lamp to (try to) max y-sort work better with it. We'll need to revisit this once we have final player dimensions / lamp location etc
- Fixes a bug where starting a game with a previously muted audio bus didn't re-mute it
- Fix a bug where we were treating `is_playing` (a function pointer) as a boolean/property causing a bug in handling Character animation state
- Added some guard conditions on the editor-run version of Devin bc i was getting errors for ... idk why actually. But this seems to solve it 🤷 
- some changes to the project rendering options so that we have pixel-perfect textures and don't get tearing as we move the camera around