Adds pushabel boxes (#24)
## Summary

This adds a bad implementation of a box that can be pushed/pulled in the four cardinal directions. I'll talk about why it's bad, why that's what is getting merged, and what you (you!) can do about it.

## New bits
`MoveableBlock`&mdash;A scene that can be detected by a Character's interaction sensor. When interacted with the character will lock into the current facing and constrain their movement into two directions to push or pull the object in targeted block.

`CharacterTarget`&mdash;type safe container for all the things that a character might interact with. In a good type system this would be a sum type. I went the route of having a container type because the alternative was one variable per type (currently MovableBlock and Interactable) which meant that there would have to be some kind of precedence logic or making sure to evict all previous target objects. That seemed prone to error and at the end of the day you basically recreate Character Target but with a less consistent interface.

`Character.SensorSet`&mdash;Instead of using a single sensor for all interactions detection we now have essentially a large and small (near/far) sensor. The small is used for things that require physical contact to look correct (currently just MoveableBlocks) and the larger sensor is everything else that doesn't look weird to interact with with a little more distance. Over time we may change the things that each is responsible for managing. Other than that they both have area enter/exit code and the logic is the same of having only one targeted item being interactable at a time.

`Enums.TargetType`&mdash;What does the character have in their sights?

`Enums.MoveMode`&mdash;walk/sprint/push_pull. This is the type of movement the character is currently doing if they move (e.g. velocity 0 can be a walk, sprint, or push_pull).

## How it works: Physics, Joints, Direct manipulation, oh my!
So there are two "reasonable" ways to handle object movement
- 100% code
- Use the force, Devin! (physics simulation)

I'm going to tell how this works but first I'll tell a story on how we got here:

### Story Time!

I'm using a `RigidBody2D` for movable blocks because we need collision and they _do not like_ their position being directly changed. That's fine, I'll just use

1. `move_and_slide`,
2. `get_slide_collision`, and
3.  `apply_central_force`

1 + 2 will detect if you're running into a `MoveableBlock` and 3 will move that block with you. Great, works perfectly!

Going backwards is harder! because pulling a block doesn't generate a collision I need to track the `_target` and then I can just apply a force in the opposite direction.

Well shit that doesn't work because it's evidently (obvious in hindsight) hard to determine the necessary force that needs to be applied in order to match the current rate that Devin is moving without doing _math_.

> In hindsight I should have just sat down and done some math to come up with a formula but I didn't.

So then I switched to using a `PinJoin2D`. A joint is basically a constraint that says "these two nodes should have a specific relationship to each other as they move." This means I can attach the `_target` when push/pull starts and then it'll all _"just work"_.

It didn't. Pulling using a joint was great but when pushing it was possible to start too close to the object and then collision would make movement _extremely_ slow if for reasons. The joint has a "disable collision" mode which could work but then Devin was able to walk straight through it when the block hit an obstacle.

I took a brief detour through direct manipulation which does work with the physics simulation even with rigid bodies when `.freeze` is true. I forget why that didn't exactly work but it wasn't great.

### Actual solution

I ended up using both `apply_central_farce` and `PinJoint2D` because I just wanted this to be done and to move on for now. I'll talk about possible improvements later.

When you press "interact" with a `MoveableBlock` focused `Character` swaps to `PUSH_PULL` move mode and the `MoveableBlock` is unfrozen which means we can safely apply forces . At this point it will snapshot the `_facing` and `_push_direction`; these will not change until the character leaves PUSH_PULL.

When taking input in PUSH_PULL the character `_impulse` gets constrained to a single axis (x or y) based on then previously cached `_push_direction`.

If the movement indicated by player input is aligned to `_push_direction` we test for collisions in `_physics_process` and then `apply_central_force` to the target. Pushing works great.

If the movement indicated by player input is _not_ the same as `_push_direction` we attach the object to a `PinJoint2D` and reduce player speed, this means as the player moves away the move more slowly (because "pulling") and the block follows along.

## PROBLEMS

This (mostly) works but there are distinct issues that we'll need to follow up on.

~~1. Because we use the interaction sensor and that thing is _huge_ relative to the size of things we want to detect and spacing--i.e., we can pull from a body width away--we'll probably want to wire up a specific sensor for push/pull operations~~
2. The whole "use two systems for push/pull" is not great and splits the push/pull logic between `_physics_process` and implicit `PinJoint2D` stuff. A custom joint, doing the physics math correctly, or direct position manipulation are all likely more optimal if we can get a single approach working for both sides. If we go this route position approach will be much easier with the cached `_push_direction` and `_is_push` methods to work with than when I first took a shot at it, just make sure to leave the bodies `.freeze=true`.
3. `_unhandled_input` has way too much going on now, we should strive to get that refactored and more approachable