Inventorty/Items System (#23)
v0 Design of the Items and Inventory systems for use in Stubs

InventoryManager will be used to store all inventories across the game. Containers or Objects that require inventories can be attached and registered via an ID for that object. We should ensure that all IDs are unique to avoid collision/overlapping.
Methods

InventoryManager#register - register a unique id for an inventory
InventoryManager#get - get an inventory via id anywhere

There's some basic Inventory UI Stubs that can be used for testing, very basic just display items on screen.

Base Item Stub Scene to create an plop down items in the world via @tool support. This means that now all you have to do to create new items is to setup the resources correctly. Items should have an associated unique ID so they can identified for other systems in the game (like quests). Right this is manual work where each resource file should be incremented by 1 every time you add a new item. I am unsure of this is something we will be able to automate going forward, or the implications of this if multiple teams are creating items so we will have to think about this a little bit more. For now it should work as is.

Use the Item Resource file to create a new item via the Editor. (We might want to add a Icon to this so it looks nicer in the Editor)

InteractableAction to pick up items in the world and drop them into an inventory. Right now this only really applies to the Player/Character. Plausible that this will change in the future, should be okay to use for now. There shouldn't be a need to create these manually outside of the ItemStubScene itself so it should be easy drop in replacement.

There is no way to currently remove items from the inventory. This needs some more internal discussion of how we want inventories to actually function so that we can handle any edge cases with ItemStack. i.e. dropping one of an item, dropping all of an item, if we want to split items

Inventories expose some signals to help understand when an inventory has picked up an item. Useful for things like a quest item being picked up to trigger something. You can hook into these by getting the objects unique ID and attaching to it's signal

signal inventory_updated(inventory: Inventory) // no promises this stays, only used for inventory ui stuff
signal inventory_item_inserted(item: ItemStack)

I've also edited Inventory/ItemStack/Item _to_string() to show an easier string representation of the entire inventory. Useful if you just want to debug quickly to see contents/how an inventory looks via console
Item Properties

id: int
name: String
description: String
stackable: bool
stack_size: int
icon: Texture2D