[minor] Rename Interactable<Action> and add reactivity docs (#55)
this is part of #36

The bulk of docs for this are included in the tgo-reactivity.md file that addresses why we're making this change in detail. A tl;dr follows though:

o We're introducing a few paths to let level designers create actions dependent on world state
o We already have one approach to this which is the Interactable scene that lets you attach actions that should be taken when the player interacts with some object
o We want to reuse all those "actions" for Interactable and all other reactive systems

...and so the first step is to rename things such that they're not semantically coupled to the Interactable scene

Merging this as a minor change since there are no functional changes and it's just a rename.

Caveat: There is one functional change which is that Interactable will now warn if it isn't configured to the correct collision layer.