Introduce Condition templates to Switches (#143)
## Issues
- closes #68
- closes #116 

This is an extension to the existing Switch configuration panel in the
addon.

Before this change you could configure some visual parameters and there
were a couple of preset templates for how the switch behaved when you
stepped on/off, i.e., what Effects fired on (de)activation.

This adds a new set of configuration templates that add **conditions**
that must be met for the switch to activate. We start with four
templates:


![image](https://github.com/user-attachments/assets/a53b6bea-035d-45cd-9a7b-436928a13126)

They do pretty much what the name suggests.

## Implementation Details

This isn't fancy but it is complicated (or at least busy). The bulk of
the work happens in `ObjectsHelperSwitchDetails.gd` where we introduce a
bunch of new configuration groups -- basically one for each condition
template because there isn't a lot of reuse that we can do. Here are the
groups all at once:


![image](https://github.com/user-attachments/assets/8834847d-ea23-4a64-8dce-f524d0dd758e)

And the corresponding scene tree is about what you'd expect


![image](https://github.com/user-attachments/assets/5fe8b636-9116-4dc3-ac64-cbcd698d6919)

You can see that each of the four condition templates have their own
configuration group that gets hidden/shown when the template is selected
from the dropdown. There are enough moving parts that it's almost worth
factoring this out into its own scene (or at least into its own code
file) but they're all orthogonal from each other and follow enough of a
pattern that it didn't feel worth the lift (yet).

Logic basically flows like so:

1. user selects a condition template
2. show relevant configuration groups, hide others
3. user makes changes then places the switch
4. build checks the condition template dropdown value and pulls the
necessary values for the switch being added

Steps 1 + 2 repeat for templates that are themselves variable (variable
id + value & quest id + state)

## New tool functions
Item gets some static functions that we use to reduce a bit of code
duplication. These should explicitly not be run from within the game and
have `tool_` prefixes and don't work if run outside the context of the
editor.

- Item.tool_all_ids() -> walks the item directory and gets all ids that
it can find in an item resource
- Item.tool_from_id() -> walks the files in the item directory until it
finds the item by the given id -- asserts if a bad ID is passed in (idk
if this crashes the editor tbh, maybe i should check that)
- QuestManager.tool_all_ids() -> returns all quest ids
- ObjectsHelpelSwitchDetails._ersatz_dialogic_get_var - this doesn't
_quite_ belong in this list but I basically had to reimplement Dialogic
VAR lookup code because we can't use it from the editor (because it
doesn't load them until game runtime). Anyway, this will get us the
default value and variable type. It doesn't safeguard at all for bad
variable ids because we're getting the path from a dropdown that pulls
them from the same source we're checking.

I briefly considered trying to aggregate these into a ToolScripts class
or something but it felt weird so for now they live attached to their
most relevant game object

## Misc changes
- Because I noticed this while testing switches / quest state effects I
made it so that a quest won't activate the _next_ quest in a phased
series unless the phase parent is active. Normally we shouldn't be in a
position where a phase quest is active without the parent _but_ it's
weird if we are and it automatically progresses imo
- QuestStateCondition is a new condition that evaluates to true when a
specified quest is in a specific state. We currently only support EXISTS
and EQ as state checks.

---------

Co-authored-by: envy <this-is-envy@users.noreply.github.com>