Rewrite character controller: Pile of shit -> State Machine (#152)
Okay, so this is meaningfully larger than it should be. If i was going
to factor this review correctly it'd come in three pieces:

1. the ControllerBase + InputController
2. the StateMachine
3. the Character rewrite to utilize the StateMachine

But that isn't the world we live in and I'm cutting it like this to
prioritize merge speed rather than a potential code review should
anybody decide to pick it up. Even so we can discuss the changes in
those units as a way to think about the review.

## ControllerBase / InputController

This is basically just an abstraction over how we get input into the
state machine. I wanted to be able to control it via behavioral agent so
that (in theory) we could have AgentController that provides movement
vector from a nav agent / behavior tree / etc. IDK if we'll ever need it
so this may be over engineering but it's in place and it works.

The theory is that we provide enough of an interface to do whatever we
need in `ControllerBase` and then the default Player-driven controller
mapping happens in `InputController`. Specifically we pass in a list of
actions (in the InputMap sense) that map to directional input and then a
list that we audit on each keypress to see their pressed state.


## StateMachine fundamentals

This is _preeeeety_ standard as far as state machines go. If you're not
familiar with the concept [here's some
writing](https://gameprogrammingpatterns.com/state.html) (Also that book
is _extremely solid_, highly recommends). As you might imagine there is
hours of content [on
youtube](https://www.youtube.com/watch?v=oqFbZoA2lnU) as well.

The somewhat interesting variation here is that each state can have args
passed in as part of the transition request. I use this to basically set
externally available information that might not be trivial to recompute
or that might change over time but which needs to be checkpointed for
the state execution. If you think of state execution as a function body
the context passed in would be arguments. These are passed via
`queue_state_change` function which is what we use to initiate a state
transfer.

In order to keep these state args "typesafe" in leiu of generics and
reliably formed the convention is that there is a static function
(prevents accidental reuse of previous instance state) on states who
receive args of `mk_args` that is static

Similarly a state machine is setup initially via `setup` and can take
arguments which might be super commonly needed for all states. An
example n our case is the `CharacterBody2D` and `AnimatedSprite2D` which
we cache in `CharacterState` since so many of the states operate on
those objects.

## Character Rewrite

The last big chunk was converting the `Character` logic to a series of
states. Most are pretty straight forward except for PushPull which has
to account for the character and the target object.

- Idle: just standing around, stops the current animation on enter; has
exits to Interact and Walk
- Walk: converts controller vector into movement; has exits to idle
(when vector == (0,0)) and interact (when you press interact)
- Interact: transitory state to interacting with something; exits to
idle (interaction complete)
- PushPull: this state gets initialized with a push target and a push
direction -- from there movement is constrained to an axis and depending
on direction the target or actor is moved with the other party moved a
corresponding amount; has exit states to idle

Closes: #46 
Closes: #92

---------

Co-authored-by: this-is-envy <this-is-envy@users.noreply.github.com>