Add sprint state and depletable stamina (#160)
Adds a stat block on characters and default stat set for Devin.
Currently includes only stamina but does cover generalized save/load
mechanics for arbitrary stats and a debug UI for tracking their values.

I added a (somewhat jank, c.f. #159) sprint state so that we have a
means to watch it deplete. Nothing wired in for dialogue or EOD reset
yet.

---------

Co-authored-by: envy <this-is-envy@users.noreply.github.com>