Quest graph utility (#164)
## Highlevel
Another larger than ideal PR but it is what it is, lol.

At the highest level this adds a new main-panel plugin for managing
quest state and reporting on configuration errors.

The quest visualization is built on top of
[GraphEdit](https://docs.godotengine.org/en/4.2/classes/class_graphedit.html);
we subclass it into QuestGraphEdit add logic to handle refresh, drag &
drop connections, etc. The quest representation is done via QuestNode
which acts as a container for a Quest objects and mostly knows how to
construct a
[GraphNode](https://docs.godotengine.org/en/4.2/classes/class_graphnode.html#class-graphnode)
to represent it.

Linting is done by the Quest and QuestCondition resources and the
results get rolled up into the LintReport with each quest getting its
own LintNodeReport (c.f. lint_node_report.gd, by the time I was done I
didn't have it in me to make the file naming consistent).

You can drag a phase or next quest output to an empty spot and it'll
create a new quest with the provided ID. This creates the `.tres` file
in Utils.QUEST_DIR and will wire up the corresponding source quest
correctly.

Changes made to quests / wiring are _not_ automaticly saved so a good
old Ctrl-S will be needed to make your edits real. Layout is saved
separately via Save Layout button.

## Screenshots

Sample quest wiring + errors + selected quest open for editing in the
far inspector panel


![image](https://github.com/user-attachments/assets/6e0c4690-f928-415d-bf32-d54bea96fcd3)

Creating a new quest via drag


![new-quest](https://github.com/user-attachments/assets/d7517b6e-a362-444f-80a0-737a66040362)

## Bugs / Future work

- We don't currently have any linting on effects
- Global linting (and general quest id error handling) could/should be
better
- We currently re-lint everything on quest updates, that is probably a
bit dramatic
- Add a QuestNode context menu or in-node button to add phases
- Global linter should check for cycles
- Layout loading is ... special and doesn't seem to fully work for a
first load; maybe rollback the "and then rearrange things that aren't
selected" bit? Also scroll_offset doesn't seem to be respected on first
display
- Dropped in quests don't quite position correctly; possibly a
scroll_offset thing?
- There is a bug that attempting to unlink a quest that was just dropped
it can't be found in the source quest's next list. It *does* seem to be
there in the data model though so idk what's up

---------

Co-authored-by: envy <this-is-envy@users.noreply.github.com>