chore: tidy up the docker bits
This commit does a load of hygiene on the Dockerfile, using smaller base
images, reducing the number of steps, removing pointless commands, etc.
It also keeps the build process local to the support folder so now the
image can be built like so:
- `docker build -t tgo:check support`

There is one breaking change is the mount point, this is now just
`/repo`.