chore: tidy up the docker bits (#193)
This commit does a load of hygiene on the Dockerfile, using smaller base
images, reducing the number of steps, removing pointless commands, etc.
It also keeps the build process local to the support folder so now the
image can be built like so:
- `docker build -t tgo-gd support`

There is one breaking change which is that the folder is not lowercase
because capital folders in a unix environment are gross.