Torch-as-gear; equipment; InputController fix (#190)
Closes: #37, #74, #75, #176

## Updated Processes, high level summary

### (New) Equipping & using gear
> Currently the only way to equip something is through the debug UI.

When a character equips gear the process is basically:
- Bail if slot is not empty
- Bail if item isn't equippable
- For each gear_spec element in Item call `on_equip(self)`
- emit `equipment_changed(id)`

Unequipping is the natural inverse operation and operates on a gear
slot, not item.
- Bail if nothing equipped in the specified slot
- Erase the old slot entry
- For each gear_spec element of the old Item call `on_remove(self)`
- emit `equipment_changed(self)`

Using an item operates on a gear slot, e.g., you will "Use LEFT_ITEM"
not "Use TORCH"
- Bail if nothing equipped in slot
- For each gear_spec element of the item call `on_use(self)`

Examples of how the Item + GearSpec resources combine to create
functionality can be seen in `Torch.tres` + `simple_torch.gd`.

### Save/load
**Save**  
When we initiate a game save we now walk the equipped items creating a
`Map[GearSlot, Variant]`. ATM that Variant is always `Array[Variant]`
because of how we handle gear_spec. This all gets shoved into the player
object in the TGO.meta file.

**Load**  
Loading is a bit surprising from equipment. Because we don't know if
we're loading from an new-game state or existing game we have to handle
potentially already equipped items. We can't simple clear the _equipment
map because that might miss any unequip cleanup.

For that reason the process of loading a game / character gear is:
- For each currently filled slot run unequip(slot) to unwind any gear
related changes
- For each slot info in the save file load the item that we find there
- Call item.restore_state on the newly equipped item with the saved data

### InputController
The previous impl here had major problems with events registering at
'just' pressed for _many_ consecutive frames. We now only store that
state for a single frame which makes this a non-issue.

This means it's technically possible to miss the JUST state of an input
but in practice we never should since `_process` and friends are
explicitly called once per frame. More troubling will be if framerate
drops and/or we get multiple `_unhandled_input` calls per frame which is
totally possible. In this case we _will_ see multiple times where
`JUST_X` will return true however _mostly_ even if that happens the
fallout should be fine.

We'll really know if this is a problem when we start doing more UI work.
## Detailed Technical Changes
- Item changes:
  - now include `gear_spec` attribute which defines what the item does
- new methods `save_state` and `restore_state` which get called during
game save & load in the event that an item has state that needs to be
persisted
- Characters changes:
- added `equipment` dictionary that maps a `GearSlot` to an `Item`
resource
  - new methods to support items:
- `_save_gear` and `_load_gear` to handle wrangling the gear into the
save file
- `equip`, `unequip`, `use_item` wrap logic that handles interaction
with gear
- `_process` now watches for `{LEFT|RIGHT}_ITEM` input and activates the
corresponding item... unless the current state is input_exclusive (see
StateMachine section below)
- new `equipment_changed` signal; currently very coarse as I only needed
it to update the debug UI rn. I could imagine a future where we signal
additionally with the slot changing etc
- GearSpec:
  - New resource that defines an interface for equipment behavior
- `on_equip`, `on_remove` both take the character equipping the item and
can be used to make any changes as needed, e.g., the `SimpleTorch`
adds/removes the torch scene to the character holding it
- `save_state`, `load_state` gets called during game state save &
restoration
  - `on_use` gets called when the user of the equipment activates it
- Note that GearSpec is a `Resource` so there is _one_ instance of it
for all
uses. That means the methods above take in a reference to the character
acting and _should not_ mutate any state on the GearSpec.
- StateMachine:
- States now have an attribute `input_exclusive` that acts as a hint to
nodes executing a StateMachine that the current state should not route
any input to ancillary behavior. This is used to prevent item activation
while in menus. It is a kludge but works for now
- InputController
- `ActionState` now tracks the frame that it entered a given state; this
us used to power `JUST_{PRESSED|RELEASED}`


## Still TODO / New work:
- May need a universal ID for items so that we can track _which_
inventory item is equipped / prevent the same item from being equipped
in multiple slots (alternatively: add `active_slot: ?GearSlot` on Item
scene?)
- Item should not be able to be equipped in more than one slot
- The Item / GearSpec process currently handles saving state by assuming
that the spec array is stable; this could seriously fuck things up if we
change item configurations. **Explicitly we should expect this to break
in weird/confusing ways if we don't change it.** A change probably looks
like some UUID for gear and converting the item state into a
`Map[ID:string, Variant]` instead of the current `Array[Variant]`.
- If we want to make some kind of store available to items for state
then we handle that as an `ItemStateStore` which handles persistence and
contains a `Map[Character, Map[Item, Variant]]` or something.
- Torch does not modulate its energy levels based on time of day /
global modulate so we currently overexpose to varying degrees during
non-night times
- Character is kinda in-the-weeds of Item/GearSpec for
equip/unequip/use. We should had that behind the Item abstraction
- Equipment save data is currently caching the resource path as the
first element of an `Array[Variant]`, it's probably better to have that
a class so that we can make it type safe, less reliant on knowing the
convention
- We might consider not having EQUIPPABLE be a full item type and rather
infer that from the presence of non-empty GearSpec or simple `can_equip`
attribute

## Sample
Torch working, toggle on/off, equipment debug UI, persist state through
save and load:


![lights](https://github.com/user-attachments/assets/6f6ef810-dda9-4cec-b16d-91068eeced48)

---------

Co-authored-by: envy <this-is-envy@users.noreply.github.com>